+++ src/Game/EchoStateNetwork/Network.py
