+++ src/Game/PlayerNeural.py
@@ -1,7 +1,7 @@
 #PlayerNeural.py
 
-from PlayerBase import PlayerBase
-from Network import Network
+from EchoStateNetwork.PlayerBase import PlayerBase
+from EchoStateNetwork.Network import Network
 
 class PlayerNeural(PlayerBase):
 	def __init__(self):
